Extracting and moving around code a bit.
