+++ furnace-gcp/commands/create_test.go
@@ -9,24 +9,6 @@ import (
 	dm "google.golang.org/api/deploymentmanager/v2"
 )
 
-type MockDeploymentService struct {
-	insert *dm.DeploymentsInsertCall
-	delete *dm.DeploymentsDeleteCall
-	get    *dm.DeploymentsGetCall
-}
-
-func (m *MockDeploymentService) Insert(project string, deployment *dm.Deployment) *dm.DeploymentsInsertCall {
-	return m.insert
-}
-
-func (m *MockDeploymentService) Delete(project string, deployment string) *dm.DeploymentsDeleteCall {
-	return m.delete
-}
-
-func (m *MockDeploymentService) Get(project string, deployment string) *dm.DeploymentsGetCall {
-	return m.get
-}
-
 func TestExecute(t *testing.T) {
 	expectedDeployments := &dm.Deployment{
 		Description: "",
